LayoutTests/http/tests/navigation/post-frames-goback1-expected.txt

Tests that a POST targetted to a frame is handled correctly when navigating away and coming back to the page.




--------
Frame: 'target-frame'
--------
This page was requested with the HTTP method POST.

Parameters:

the-input = input value goes here

============== Back Forward List ==============
curr->  http://127.0.0.1:8000/navigation/post-frames-goback1.html
            http://127.0.0.1:8000/navigation/resources/form-target.pl (in frame "target-frame")
        http://127.0.0.1:8000/navigation/resources/go-back.html
===============================================
